[+] added patterned file search to fileTools

+++ tests/FileToolsTest.py
@@ -14,11 +14,17 @@ class FileToolsTest(unittest.TestCase):
         readed = FileTools.getContents(TEST_FILE)
         self.assertEqual(readed, TEST_DATA)
 
+    def test_getDirContentsRec(self):
+        res = FileTools.getDirContentsRec("./")
+        self.assertTrue("./testFiles/test1" in res)
+        self.assertTrue("./testFiles/test2" in res)
+        res = FileTools.getDirContentsRec("./", "*2")
+        self.assertFalse("./testFiles/test1" in res)
+        self.assertTrue("./testFiles/test2" in res)
+
     def tearDown(self):
-        os.remove(TEST_FILE)
+        if os.path.exists(TEST_FILE):
+            os.remove(TEST_FILE)
 
 if __name__ == "__main__":
     unittest.main()
-
-
-
